# prune unneeded gl calls