This PR removes the glClear and glDisable calls to help shave some additional ms' off of the GPU render stack on lower end iOS hardware such as the iPad Air 1.