@mrjschulte 2019-09-02
Together with #1176 these two teaks result in the following performance metrics testing on the Genesis Core with Dragons Fury and using Multi-threaded GL:

**Before:**
Utilization
Tiler - 37%
Renderer - 34% 
Device - 37%
**Frame Time**
CPU - 0ms
GPU  - **6.1ms**

**After** 
Utilization
Tiler - 20%
Renderer - 16% 
Device - 20%
**Frame Time**
CPU - 0ms
GPU  - **3.3ms**