@JoeMatt 2019-09-09
Has this been tested to not have GFX glitches on all cores?

Specifically the off thread GL based cores like N64, Reicast.