@braindx 2019-09-09
It would probably be better to leave the glClear calls in development builds as it can aid in debugging new cores and video modes/resolutions, but I would think it’s probably fine to skip those generally.

However, I would think it’s possible that skipping the glDisable calls causes issues with certain cores and titles that use OpenGL natively, e.g. certain N64 games, as they may have set up OpenGL state that prevents the subsequent render call from drawing correctly or at all.