@JoeMatt 2019-09-09
Sounds like there may be 2 modifications to consider then,

1. #ifdef DEBUG wrap the glClear calls
2. Wrap the glDisable calls in either an `if core.rendersToOpenGL`, or two versions of the function with and without and use a function pointer at Core startup to set the function pointer. This might be an over-optimization to avoid 1 if statement, though would perhaps allow future develop for native and non-native gl render path ways.