@mrjschulte 2019-09-09
Fine with whatever you guys think is best on how to wrap these tweaks in. 

I know the midterm plan is to move Pv to a native Metal renderer at some point but these tweaks get the lower pieces of the ARM SKU's the full CRT shader in landscape mode.